Move state caches to chain/stateCache
