+++ packages/lodestar/src/chain/regen/regen.ts
@@ -4,8 +4,7 @@ import {computeEpochAtSlot, computeStartSlotAtEpoch} from "@chainsafe/lodestar-b
 import {IForkChoice} from "@chainsafe/lodestar-fork-choice";
 
 import {ITreeStateContext} from "../interface";
-import {StateContextCache} from "../stateContextCache";
-import {CheckpointStateCache} from "../stateContextCheckpointsCache";
+import {CheckpointStateCache, StateContextCache} from "../stateCache";
 import {ChainEventEmitter} from "../emitter";
 import {IBeaconDb} from "../../db";
 import {processSlotsByCheckpoint, runStateTransition} from "../blocks/stateTransition";
