+++ packages/lodestar/src/chain/stateCache/index.ts
@@ -1,0 +1,2 @@
+export * from "./stateContextCache";
+export * from "./stateContextCheckpointsCache";
