+++ packages/lodestar/test/utils/mocks/chain/chain.ts
@@ -20,8 +20,7 @@ import {IForkChoice} from "@chainsafe/lodestar-fork-choice";
 import {ChainEventEmitter, IBeaconChain, ITreeStateContext} from "../../../../src/chain";
 import {IBeaconClock} from "../../../../src/chain/clock/interface";
 import {generateEmptySignedBlock} from "../../block";
-import {StateContextCache} from "../../../../src/chain/stateContextCache";
-import {CheckpointStateCache} from "../../../../src/chain/stateContextCheckpointsCache";
+import {CheckpointStateCache, StateContextCache} from "../../../../src/chain/stateCache";
 import {LocalClock} from "../../../../src/chain/clock";
 import {IStateRegenerator, StateRegenerator} from "../../../../src/chain/regen";
 import {StubbedBeaconDb} from "../../stub";
